requirements.txt
httpx==0.27.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
loguru==0.7.2
pyyaml==6.0.1
curl-cffi==0.7.3
# Browser automation for payment (bypasses anti-bot)
playwright==1.41.0
# Telegram bot
python-telegram-bot==21.0
# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
